# shared/schemas/api.openapi.yaml
openapi: 3.0.3
info:
  title: Family Trivia API
  description: Backend API for the Family Trivia game
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local SAM development
paths:
...
  /generate-question:
    post:
      summary: Generate a trivia question
      operationId: generateQuestion
      tags: [Trivia Questions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, mode]
              properties:
                userId:
                  type: string
                mode:
                  $ref: "#/components/schemas/GenerateMode"
                category:
                  type: string
                difficulty:
                  $ref: "#/components/schemas/DifficultyLevel"
                customPrompt:
                  type: string
                isCatchingUp:
                  type: boolean
                withProgress:
                  type: boolean
      responses:
        "200":
          description: Generated question with optional progress
          content:
            application/json:
              schema:
                type: object
                properties:
                  question:
                    $ref: "#/components/schemas/Question"
                  messages:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProgressMessage"
                  factCheckResult:
                    type: object
                  retriesNeeded:
                    type: integer
...
  /generate-question-options:
    post:
      summary: Generate question options for daily fact
      operationId: generateQuestionOptions
      tags: [Daily Facts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: string
                theme:
                  type: string
                regenerate:
                  type: boolean
      responses:
        "200":
          description: Generated options
          content:
            application/json:
              schema:
                type: object
                properties:
                  options:
                    type: array
                    items:
                      type: object
                      properties:
                        question:
                          type: string
                        theme:
                          type: string
                  weekSummaryContext:
                    type: string
  /recent-shared-questions:
    get:
      summary: Get recent shared questions
      operationId: getRecentSharedQuestions
      tags: [Daily Facts]
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Recent shared questions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    question:
                      type: string
                    date:
                      type: string
...
components:
  schemas:
...
    DifficultyLevel:
      type: string
      enum: [easy, normal, hard]
    GenerateMode:
      type: string
      enum: [personal, history, group, category, custom]
    Question:
      type: object
      required: [question, choices, answer]
      properties:
        question:
          type: string
        choices:
          type: array
          items:
            type: string
        answer:
          type: string
        category:
          type: string
        difficulty:
          $ref: "#/components/schemas/DifficultyLevel"
        pointMultiplier:
          type: number
        mode:
          type: string
        type:
          type: string
        source:
          type: string
        sourceUrl:
          type: string
        externalId:
          type: string
    ProgressMessage:
      type: object
      required: [type, message, timestamp]
      properties:
        type:
          type: string
          enum: [info, warning, success, checking]
        message:
          type: string
        timestamp:
          type: number
        detail:
          type: string
        duration:
          type: number
...